chore(app): separate epoch change executor logic
